module6-stream-processing/risingwave/pyproject.toml
[project]
name = ""
version = "1.0"
description = "Rising Wave Workshop"
license = {text = "CC-BY-SA-4.0"}
readme = "README.md"
authors = [
    {name = "Bruno Oliveira", email = "bruno.ariev@gmail.com"},
]

requires-python = ">=3.10,<3.12"

dependencies = [
    "fastapi>=0.110.0",
    "polars>=0.20.15,<1.0",
    "pandas>=2.2.1,<3.0",
    "pyarrow>=15.0.1,<16.0",
    "psycopg2-binary==2.9.9",
    "confluent-kafka>=2.3.0,<3.0",
    "uvicorn>=0.28.0",
]

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.3.2",
]

[tool.ruff]
line-length = 100
lint.select = ["F", "E", "W", "I001"]
